% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTime.R
\name{date.to.DOY}
\alias{date.to.DOY}
\title{Date conversion}
\usage{
date.to.DOY(dates, format = "dd/mm/yyyy")
}
\arguments{
\item{dates}{Vector with dates to convert.}

\item{format}{Format of the date (support DD/MM/YYYY MM/DD/YYYY and YYYY/MM/DD).}
}
\value{
Return a vector containing the corresponding DOY.
}
\description{
Convert dates from the DD/MM/YYYY format to day of the year (DOY)
}
\examples{
dates = c("01/01/2000", "03/03/2000", "03/03/1999")
date.to.DOY(dates=dates)
}
