% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTime.R
\name{datetime}
\alias{datetime}
\title{Time & dates conversion}
\usage{
datetime(dates, Time)
}
\arguments{
\item{dates}{Vector with dates in the DOY format.}

\item{Time}{Vector with time}
}
\value{
Return a vector containing DOY and time as a single numerical variable
}
\description{
Convert DOY and time into a single numerical variable
}
\details{
time vector should be numerical (e.g. as outputed by the time.to.cont function)
}
\examples{
dates = c(102,102,102,102,103,103,103,103)
Time = c(22, 22.5, 23, 23.5, 0, 0.5, 1, 1.5)
datetime(dates=dates, Time=Time)
}
