% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaflowCalc.R
\docType{data}
\name{SpFl}
\alias{SpFl}
\title{Sap flow dataset}
\format{An object of class \code{data.frame} with 432 rows and 4 columns.}
\usage{
SpFl
}
\description{
Exemple dataset exemple for the TDPanalysis package
}
\details{
"DATE" is dates in dd/mm/yyyy format. "TIME" is time in hh:mm:ss format, "ID" is sub-groups and "tension" is the measured tension from the TDP probe.
}
\keyword{datasets}
