% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTime.R
\name{timecont}
\alias{timecont}
\title{Time conversion}
\usage{
timecont(Time, sep = ":")
}
\arguments{
\item{Time}{Vector with time to convert.}

\item{sep}{Character element containing regular expression(s) to use to splitting.}
}
\value{
Return a vector containing the corresponding time.
}
\description{
Convert time from the HH:MM:SS format to a numerical
}
\details{
time vector should be in the HH:MM:SS format.
}
\examples{
Time = c("14:30:00", "20:45:00", "05:00:00")
timecont(Time=Time)
}
